Revert ProductDetailNavigation changes
This reverts commit: 73c7cea721487bc09f38da4750ef4d6f79ec8e8e, and f4adca43, and 7ea36dd8f0f8883b1f027a3b464f0f54cd52d9bc.
